apps/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.35
asyncpg==0.30.0
redis==5.2.0
nats-py==2.9.0
boto3==1.35.36
minio==7.2.9
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
httpx==0.27.2
aiofiles==24.1.0
psycopg2-binary==2.9.9
alembic==1.13.0
python-json-logger==2.0.7

